+++ lizard_damage/tools.py
@@ -9,4 +9,7 @@ def version():
     from pkginfo.installed import Installed
     import lizard_damage
     installed = Installed(lizard_damage)
-    return 'Versie oktober 2012 (%s)' % (installed.version)
+    if installed:
+        return 'Versie oktober 2012 (%s)' % (installed.version)
+    else:
+        return 'Versie oktober 2012'